refactor: move output package to internal
