+++ architecture.md
@@ -105,7 +105,7 @@ The directory map below folds the high-level module relationships into the packa
 | ------------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `website/`                     | Documentation site and Playground UI                                                                                                        | Uses `packages/rslint-wasm` to run browser linting and `packages/rslint-api` to decode encoded source files; Playground lint requests ultimately reach `internal/linter`, and inspect requests reach `internal/inspector` through `internal/api`                                       |
 | `cmd/rslint/`                  | Main Go binary entry point with CLI, API, and LSP modes                                                                                     | JS/TS config CLI path is `internal/config/discovery -> internal/config -> cmd/rslint/programs.go -> internal/linter`; JSON config starts at `internal/config`. `--api` is consumed by `packages/rslint` and `packages/rslint-wasm`; `--lsp` is consumed by `packages/vscode-extension` |
-| `cmd/rslint/internal/output/`  | CLI report model, summary, colors, and stdout formatters                                                                                    | Consumes final sorted `internal/rule` diagnostics from the CLI pipeline and renders `default`, `jsonline`, `github`, or `gitlab`; it is intentionally private to `cmd/rslint` and is not shared by the structured API or LSP adapters                                                  |
+| `internal/output/`             | Report model, summary, colors, and stdout formatters                                                                                        | Consumes final sorted `internal/rule` diagnostics and renders `default`, `jsonline`, `github`, or `gitlab`; the CLI is its current consumer, while the package remains available to other repository integrations that need the same output behavior                                   |
 | `cmd/tsgo/`                    | ts-go semantic inspection/export tool                                                                                                       | Talks directly to `typescript-go` and bypasses the lint framework; consumed by `packages/tsgo` and `crates/tsgo-client`                                                                                                                                                                |
 | `internal/api/`                | stdio IPC protocol and service types for JS/WASM integration                                                                                | Shared protocol layer for `cmd/rslint --api`; used by `packages/rslint`, `packages/rslint-wasm`, `internal/linter`, and `internal/inspector`                                                                                                                                           |
 | `internal/config/`             | Configuration models, JSON loading, matching/merging, runtime ownership resolution, lint-target planning, and centralized rule registration | Owns the shared authored-global-ignore matcher consumed by both discovery phases. `RegisterAllRules()` orchestrates rule registration; `rule_registry.go` implements registry/query logic used by `cmd/rslint/programs.go` and `internal/linter`                                       |
@@ -190,7 +190,7 @@ The current parsing and linting pipeline is built on top of ts-go `Program` and 
 6. **AST Traversal**: The linter traverses each file once using a DFS walk.
 7. **Rule Execution**: Listener callbacks inspect nodes and may use syntax only or syntax plus type information.
 8. **Diagnostic Collection**: Findings are reported as `RuleDiagnostic` values, with optional fixes or suggestions.
-9. **Output Generation**: CLI builds one report from the final post-fix diagnostics and passes it to `cmd/rslint/internal/output`; API returns structured data, and LSP converts diagnostics to LSP diagnostics/code actions.
+9. **Output Generation**: CLI builds one report from the final post-fix diagnostics and passes it to `internal/output`; API returns structured data, and LSP converts diagnostics to LSP diagnostics/code actions.
 
 ### Error Recovery Strategy
