+++ internal/output/color.go
