+++ internal/output/diagnostic.go
